// src/api/generated/model/updateEmailSettingsCommand.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with OpenIddict Authentication
 * OpenAPI spec version: v1
 */

export interface UpdateEmailSettingsCommand {
  /** @nullable */
  smtpHost?: string | null;
  smtpPort?: number;
  /** @nullable */
  smtpUser?: string | null;
  /** @nullable */
  smtpPassword?: string | null;
  /** @nullable */
  fromEmail?: string | null;
  /** @nullable */
  fromName?: string | null;
}